moved generator function to root of dataprofiler dir; added test_generator.py; reverted test_data_utils and test_utils
